+++ src/app/core/config/profiles/edc-ui-profile-data.ts
@@ -1,6 +1,7 @@
 import {MDS_THEME, SOVITY_THEME} from './edc-ui-theme-data';
 import {COMMUNITY_EDITION_FEATURES} from './feature-sets/community-edition-features';
 import {ENTERPRISE_EDITION_FEATURES} from './feature-sets/enterprise-edition-features';
+import {MDS_FEATURES} from './feature-sets/mds-features';
 import {inferEdcUiProfileType} from './infer-edc-ui-profile-type';
 
 /**
@@ -22,25 +23,25 @@ export const EDC_UI_PROFILE_DATA = inferEdcUiProfileType({
   'mds-open-source': {
     ...MDS_THEME,
     routes: 'connector-ui',
-    features: new Set(['mds-fields', ...COMMUNITY_EDITION_FEATURES]),
+    features: new Set([...MDS_FEATURES, ...COMMUNITY_EDITION_FEATURES]),
   },
   'mds-hosted-by-sovity': {
     ...MDS_THEME,
     routes: 'connector-ui',
     features: new Set([
-      'mds-fields',
-      'mds-ee-basic-marketing',
+      'mds-marketing',
+      ...MDS_FEATURES,
       ...ENTERPRISE_EDITION_FEATURES,
     ]),
   },
   'mds-blue-hosted-by-sovity': {
     ...SOVITY_THEME,
     routes: 'connector-ui',
-    features: new Set(['mds-fields', ...ENTERPRISE_EDITION_FEATURES]),
+    features: new Set([...MDS_FEATURES, ...ENTERPRISE_EDITION_FEATURES]),
   },
   broker: {
     ...MDS_THEME,
     routes: 'broker-ui',
-    features: new Set(['mds-fields']),
+    features: new Set(MDS_FEATURES),
   },
 });