+++ src/app/routes/connector-ui/policy-definition-page/new-policy-dialog/new-policy-dialog-form-model.ts
@@ -16,7 +16,7 @@ export interface NewPolicyDialogFormModel {
     start: FormControl<Date | null>;
     end: FormControl<Date | null>;
   }>;
-  connectorId: FormControl<string>;
+  participantIds: FormControl<string[]>;
 }
 
 export type PolicyType =