+++ CHANGELOG.md
@@ -18,6 +18,10 @@ the detailed section referring to by linking pull requests or issues.
 
 #### Minor
 
+- MDS Connector UI: Renamed Participant IDs to MDS Connector IDs
+- Connector Restricted Policy now supports entry of multiple connectors.
+- New Transfer History Page Column: Participant ID / Connector ID
+
 #### Patch
 
 #### Deployment Migration Notes